refactor(app-bar): integrate burger menu into main app bar component
- Move burger menu logic from separate BurgerMenu component into AppBar
- Add Menu/X icons from lucide-react for toggle button
- Add isOpen state to manage mobile menu visibility
- Implement mobile menu items directly in AppBar (My Code, Account links)
- Delete burger-menu.tsx (no longer needed)
- Update SignInButton mobile styles for better layout consistency
- Change header min-height and add shadow-md for improved visual separation

This simplifies component structure by consolidating mobile menu logic.
